modifico funciones que (de)serializan pcb

+++ commons/commons/structUtiles.h
@@ -67,7 +67,7 @@ typedef struct{
 	t_list* indiceCodigo;
 //	t_intructions* indiceCodigo;
 	t_list* indiceStack;
-	int16_t exitCode;
+	int32_t exitCode;
 	uint32_t consolaFd;
 	char* etiquetas;
 	uint32_t stackPointer; // el ultimo offset
